backend/requirements.txt
affine==2.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.116.1
fiona==1.10.1
fonttools==4.59.2
geopandas==1.1.1
h11==0.16.0
idna==3.10
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
matplotlib==3.10.6
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.2
openpyxl==3.1.5
oracledb==3.3.0
packaging==25.0
pandas==2.3.2
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyogrio==0.11.1
pyparsing==3.2.3
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
pyzmq==27.0.1
rasterio==1.4.3
reportlab==4.4.3
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.2
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
wcwidth==0.2.13
